Add files via upload
This file shows the classifiers that have been trained on the data, to see which one performs best. The one that performed best was used as a classifier to predict political leaning after the two obfuscation methods. Another classifier (SVM) was also trained, but this is in a different (separate) file.